refactor: adjust spaces in repository interfaces files
