+++ domain/src/main/java/com/baghdad/domain/repository/AuthenticationRepository.kt
@@ -1,10 +1,14 @@
 package com.baghdad.domain.repository
 
-import com.baghdad.entity.User
+import com.baghdad.entity.user.User
 
 interface AuthenticationRepository {
+
     suspend fun login(userName: String, password: String)
+
     suspend fun isUserLoggedIn(): Boolean
+
     suspend fun getLoggedInUser(): User?
+
     suspend fun logOut(): Boolean
 }
