+++ domain/src/main/java/com/baghdad/domain/repository/ContinueWatchingRepository.kt
@@ -1,24 +1,24 @@
 package com.baghdad.domain.repository
 
-import com.baghdad.domain.model.ContinueWatching
-import com.baghdad.domain.model.PagedResult
+import com.baghdad.domain.model.continueWatching.UserWatchedMedia
+import com.baghdad.domain.model.pagination.PagedResult
 import kotlinx.coroutines.flow.Flow
 
 interface ContinueWatchingRepository {
     suspend fun getContinueWatching(
         page: Int,
         pageSize: Int,
-    ): PagedResult<ContinueWatching>
+    ): PagedResult<UserWatchedMedia>
 
-    suspend fun observeContinueWatching(): Flow<List<ContinueWatching>>
+    suspend fun observeContinueWatching(): Flow<List<UserWatchedMedia>>
 
     suspend fun addContinueWatching(
         contentId: Long,
         genreIds: List<Long>,
         contentImageUrl: String,
-        contentType: ContinueWatching.ContentType,
+        contentType: UserWatchedMedia.ContentType,
     )
 
-    suspend fun getAllContinueWatchingMovies(): Flow<List<ContinueWatching>>
-    suspend fun getAllContinueWatchingTvShows(): Flow<List<ContinueWatching>>
+    suspend fun getAllContinueWatchingMovies(): Flow<List<UserWatchedMedia>>
+    suspend fun getAllContinueWatchingTvShows(): Flow<List<UserWatchedMedia>>
 }
