+++ domain/src/main/java/com/baghdad/domain/repository/EpisodeRepository.kt
@@ -1,13 +1,14 @@
 package com.baghdad.domain.repository
 
-import com.baghdad.domain.model.MediaAccountStates
 import com.baghdad.entity.media.Episode
 import com.baghdad.entity.person.CastMember
 
 interface EpisodeRepository {
-    suspend fun getEpisodeDetails(tvId: Long, seasonNumber: Int, episodeNumber: Int): Episode
+
+    suspend fun getEpisodeDetails(tvShowId: Long, seasonNumber: Int, episodeNumber: Int): Episode
+
     suspend fun getEpisodeCastMembers(
-        tvId: Long,
+        tvShowId: Long,
         seasonNumber: Int,
         episodeNumber: Int
     ): List<CastMember>
@@ -23,5 +24,5 @@ interface EpisodeRepository {
         tvShowId: Long,
         seasonNumber: Int,
         episodeNumber: Int
-    ): MediaAccountStates
+    ): Boolean
 }
