+++ domain/src/main/java/com/baghdad/domain/repository/RecentlyViewedRepository.kt
@@ -4,8 +4,10 @@ import com.baghdad.domain.model.search.RecentlyViewed
 import kotlinx.coroutines.flow.Flow
 
 interface RecentlyViewedRepository {
+
     suspend fun getAllRecentlyViewed(): Flow<List<RecentlyViewed>>
 
     suspend fun deleteAllRecentlyViewed()
+
     suspend fun addRecentlyViewed(recentlyViewed: RecentlyViewed)
 }
