+++ domain/src/main/java/com/baghdad/domain/repository/SavedListRepository.kt
@@ -1,20 +1,26 @@
 package com.baghdad.domain.repository
 
-import com.baghdad.domain.model.PagedResult
+import com.baghdad.domain.model.pagination.PagedResult
 import com.baghdad.domain.model.savedList.SavedListDetails
 import com.baghdad.entity.savedList.SavedList
 
 interface SavedListRepository {
+
     suspend fun createSavedList(title: String)
+
     suspend fun getSavedLists(
         page: Int,
         pageSize: Int,
     ): PagedResult<SavedList>
+
     suspend fun addMovieToSavedList(listId: Long, movieId: Long)
+
     suspend fun getSavedListDetails(
         listId: Long,
         page: Int,
-        pageSize: Int): SavedListDetails
+        pageSize: Int
+    ): SavedListDetails
+
     suspend fun removeMovieFromSavedList(listId: Long, movieId: Long)
 
     suspend fun deleteSavedListById(listId: Long)
