+++ domain/src/main/java/com/baghdad/domain/repository/SearchRepository.kt
@@ -1,22 +1,27 @@
 package com.baghdad.domain.repository
 
-import com.baghdad.domain.model.PagedResult
-import com.baghdad.domain.model.savedList.SavableMovie
+import com.baghdad.domain.model.pagination.PagedResult
+import com.baghdad.domain.model.savedList.SavedMovie
 import com.baghdad.entity.media.TvShow
 import com.baghdad.entity.person.Actor
 import com.baghdad.entity.search.RecentSearch
 import kotlinx.coroutines.flow.Flow
 
 interface SearchRepository {
+
     fun getRecentSearches(): Flow<List<RecentSearch>>
-    suspend fun deleteRecentSearchById(id : Long)
+
+    suspend fun deleteRecentSearchById(id: Long)
+
     suspend fun deleteAllRecentSearches()
+
     suspend fun searchActorsByName(name: String, page: Int, pageSize: Int = 20): PagedResult<Actor>
+
     suspend fun searchMoviesByTitle(
         title: String,
         page: Int,
         pageSize: Int = 20,
-    ): PagedResult<SavableMovie>
+    ): PagedResult<SavedMovie>
 
     suspend fun searchTvShowsByName(
         title: String,
